+++ app/src/main/java/com/alexm/chess960/clockpack/mvp/ClockView.kt
@@ -1,8 +1,11 @@
-package com.alexm.chess960.clockpack
+package com.alexm.chess960.clockpack.mvp
 
 import com.alexm.chess960.PausePlayState
 import com.alexm.chess960.RunningClock
 
+/**
+ * Created by alexm on 04/10/2019
+ */
 interface ClockView {
     fun showCountdown(remaningTime: String, clockSel: RunningClock)
     fun enableButton1(enabled: Boolean)
