+++ app/src/main/java/com/alexm/chess960/randompos/mvp/RandomPresenter.kt
@@ -1,19 +1,12 @@
-package com.alexm.chess960.randompos
+package com.alexm.chess960.randompos.mvp
 
-import io.reactivex.rxjava3.android.schedulers.AndroidSchedulers
-import io.reactivex.rxjava3.core.Observable
+import com.alexm.chess960.randompos.RandomPos
 import io.reactivex.rxjava3.core.Observer
 import io.reactivex.rxjava3.disposables.Disposable
-import io.reactivex.rxjava3.schedulers.Schedulers
-
 
 /**
  * Created by alexm on 04/10/2019
  */
-interface RandomView {
-    fun showRandomPos(pos: RandomPos)
-}
-
 class RandomPresenter(private val logic: RandomLogic) {
 
     private var view: RandomView? = null
@@ -47,17 +40,5 @@ class RandomPresenter(private val logic: RandomLogic) {
 
             }
         })
-    }
-}
-
-class RandomLogic {
-    fun generateRandomPos(): Observable<RandomPos> {
-        return Observable.defer {
-            val pos = RandomPos()
-            pos.shuffle()
-            Observable.just(pos)
-        }
-                .subscribeOn(Schedulers.computation())
-                .observeOn(AndroidSchedulers.mainThread())
     }
 }
