+++ app/src/main/java/com/alexm/chess960/randompos/mvp/RandomView.kt
@@ -1,0 +1,10 @@
+package com.alexm.chess960.randompos.mvp
+
+import com.alexm.chess960.randompos.RandomPos
+
+/**
+ * Created by alexm on 04/10/2019
+ */
+interface RandomView {
+    fun showRandomPos(pos: RandomPos)
+}
